Fix Intern configuration
This enables easy testing using Intern.
Note that `waitSeconds` is set for the RequireJS loader because on
slow machines it can take several seconds for instrumentation on
very large libraries like jQuery to complete, and RequireJS will
treat the request as timed out even though it is actually still in
progress.